I have implemented gradient descent for both alpha and beta, but I now have a segmentation fault. Maybe it is happening in the function that calculates the beta-derivative becuae I did not call that function before.
I have found parameters for alpha and beta in Mortens notebook, so now I can try to find out what is wrong and try to get E = 3 with the Jastrow factor. I think I still have a non-zero variance for the case without the Jastrow factor, so somthing might be wrong there as well.