Some progress
Have found some optimized values and calculated some one-body densities. It takes a long time to get the optimal values for the different omegas because I have to go slowly down from one value to the next (finding the parameterrs for the omegas in between). I did suddently wonder if it should be he same values for all omegas like for the non-interacting case? Will have to ask someone about that.